# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkyBooker API",
        "description": "Flight Booking System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/airports": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "List all airports",
                "operationId": "79fcfcd17e9ea160f2d511c059520196",
                "responses": {
                    "200": {
                        "description": "List of airports"
                    }
                }
            },
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Create a new airport",
                "operationId": "17319851f20fb94dcf11209c2f34741e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "city",
                                    "country",
                                    "terminals",
                                    "status"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "AMS"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Amsterdam Schiphol"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Amsterdam"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Netherlands"
                                    },
                                    "terminals": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Airport created successfully"
                    }
                }
            }
        },
        "/api/v1/airports/{code}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get airport by code",
                "operationId": "b31a4f1d4a3615200d7276dcdae8b3ec",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Airport code (3 letters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "JFK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airport retrieved successfully"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airports"
                ],
                "summary": "Update an airport",
                "operationId": "da84d726176849b19365f79fcde5d0c4",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Airport code (3 letters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "JFK"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Name"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "New City"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "New Country"
                                    },
                                    "terminals": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "maintenance"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Airport updated successfully"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Airports"
                ],
                "summary": "Delete an airport",
                "operationId": "b80ad43db9652b4e6878a096c449f5fe",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Airport code (3 letters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "JFK"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Airport deleted successfully"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            }
        },
        "/api/v1/airports/search": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Search airports",
                "description": "Search airports by code, name, or city",
                "operationId": "ea67c96237568e9b735e52b52e11b8d0",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Airport code",
                        "schema": {
                            "type": "string",
                            "example": "JFK"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Airport name",
                        "schema": {
                            "type": "string",
                            "example": "John F"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City name",
                        "schema": {
                            "type": "string",
                            "example": "New York"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                }
            }
        },
        "/api/v1/airports/{code}/flights": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get all flights for an airport",
                "operationId": "abd5ba6a911eb7c2a5721ada80b5098b",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Airport code (3 letters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "JFK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of flights for the airport"
                    },
                    "404": {
                        "description": "Airport not found"
                    }
                }
            }
        },
...
    },
...
}